Use bounds instead of safe area (#309)
I'm not sure why, but MainView.safeAreaLayoutGuide.layoutFrame is zero in the context of React Native. However, we shouldn't try to accommodate the notch at this layer anyway. The owning view should do that, and we should make the subview match so that it just looks like we are drawing to the same view. For this, we can just use the bounds.

This change is related to #278