[iOS + XR] Fix camera rotation for portrait display orientations (#389)
This bug manifested as the transformation of screen coordinates to picking rays resulting in rays with the x/y reversed from what was expected. Ultimately the problem was that when we rotate the camera pose around the z axis for portrait and portrait upside down display orientations, the signs were swapped. Therefore, in regular portrait mode, the camera was actually upside down. There was other code that tried to account for this by flipping the UVs for both the Babylon texture and the camera texture. So the fix here is to:

1. Rotate the camera pose around the z-axis the correct direction for portrait/portrait upside down display orientations.
2. Don't reverse the UVs for the Babylon texture (note that the V is still swapped, but this is expected based on UV conventions in Metal - see https://developer.apple.com/documentation/metal/creating_and_sampling_textures).
3. Don't reverse the UVs for the camera texture.