Fix a few memory management issues in xr.mm (#456)
The main issue being fixed here is that after
id<MTLTexture> texture = [metalDevice newTextureWithDescriptor:textureDescriptor];
We need an explicit call to retain. In Cocoa, these static factory functions always call autorelease for the object being constructed, which means release will automatically be called when the containing @autoreleasepool scope exits. However, we previously didn't have an @autoreleasepool scope because we were managing the JS thread scheduling ourselves (and I didn't understand memory management patterns with Objective C and Cocoa). React Native's CallInvoker::invokeAsync does this correctly by wrapping the call to the function passed to it in an @autoreleasepool. What this means is that the @autoreleasepool is actually auto releasing the MTLTexture, so the textures are actually being released too early (I guess it usually works because the memory hasn't been overwritten yet) and then our code also explicitly releases it, so it is also double released (or rather we try to release an object that has already been released). So the proper way to deal with this is to explicitly call retain to match our explicit call to release.

I fixed up some additional things in this file now that I have a better understanding of Objective C + Cocoa memory management.

- Use a factory function to create MTLTextureDescriptors (which calls autorelease) and then just let the CallInvoker::invokeAsync scoped @autoreleasepool clean them up.
- Remove all direct usages of dealloc. This is equivalent to calling delete on a ref counted object, which is bad.
- Fix the fallout from the above (e.g. release the child MTKView after inserting it into the parent, since the parent will retain it and release it when it is removed from the children collection).