Fix texture data lifetime management issues (#480)
There were a couple of problems with our lifetime management of buffers obtained through XMLHttpRequest that contain texture data loaded asynchronously:

- The native memory was managed by the lifetime of the underlying UrlRequest, owned by the native XMLHttpRequest. If the JavaScript side asks for the ArrayBuffer of the downloaded data and holds onto it but doesn't hold onto the XMLHttpRequest, then that native memory is deleted from under us and we end up accessing deleted memory while asynchronously loading a texture (or possibly other data). The fix for this for now is to make a copy of the data into the ArrayBuffer so the JS side can own it. We could improve on this by ref counting the native memory, but we should address that with #445.
- The native code that does the async texture loading does not take a strong reference on the JS ArrayBuffer, so it could be garbage collected before we finish reading it to create a texture. The fix for this is to capture a strong ref via Napi::Persistent.

Having these fixes in place, I'm also switching back to Chakra instead of V8 for Windows as it seems to reveal these issues better (maybe a more aggressive garbage collector).

One other change I'm sneaking in to this PR is to reduce the max dump count from 2 to 1. I originally used 2 just so we would know if something unexpected is happening, but when you re-run an existing build, it re-runs on the same agent and will add a second crash dump to the dump folder which is then uploaded with the artifact. We don't want this - we just want the dump that was generated for the current run. Setting the max dump count to 1 will result in any previous dumps being deleted before the new one is written.