UrlLib/XMLHttpRequest error handling (#479)
On both Android and iOS, if a client side error (such as no internet connection) occurred, the app would crash. In the case of iOS, it was because we were throwing exceptions inside a Cocoa callback rather than propagating them throw the arcana task returned from SendAsync. In the case of Android, it was because we were not checking for Java exceptions and converting them to C++ exceptions.

I also snuck one other small change in here, which is to not copy data out of NSData into a std::vector, but rather just take a strong reference to the NSData and return a gsl::span that directly references the NSData. This just eliminates an extra data copy, which can be rather large depending on the file being downloaded.