Support updating window in ARKit XR (#487)
We were previously just passing a void* windowPtr into the XR layer (specifically for iOS/ARKit right now), but this isn't good enough since the window can change during an XR session. The change here is to instead pass in a std::function<void*()> windowProvider so the XR layer can query the current window. This happens inside GetNextFramed, which is similar to what the ARCore XR layer does (it polls the current gl surface on each GetNextFrame). This addresses BabylonJS/BabylonReactNative#106. Note the reason it was crashing before is because we immediately released the xrView and allowed the superview to manage its lifetime, but the superview (the window) can be replaced, in which case the xrView is released and we continue to try to use it. Now we retain shared ownership and don't release the xrView until we end the session.