WebXRImageTracking support for iOS (#1059)
This PR builds on top of the NativeXR and ARCore implementation of WebXR image tracking in #1032 by extending it to support ARKit on iOS as part of #619.

This is designed very similarly to the existing implementations for Planes and Meshes on ARKit as they are all built on top of ARKit anchors. The primary difference here is when constructing the image database at session start validation is offloaded to a worker thread to not block frame processing as this can block for multiple frames during XR initialization.

I have also consolidated Planes, Meshes, and Image anchors to use a single lock that covers all anchor updates.