Rename local variable `bestPixelFormat` to `devicePixelFormat`
The "best" prefix in the `bestPixelFormat` local variable name only makes sense for the current iOS implementation. It doesn't make sense on macOS since we're not selecting a "best" device. On macOS we're selecting the default device.

This change updates the local variable name so it makes sense for the current implementation on both iOS and macOS.