Fix iOS NativeCamera texture pixel format (#1125)
* Turn ARC on in Apple NativeCamera plugin CMake

* Fix ARC compile errors

* Reject unsupported pixel formats

* Set video output pixel format to same as video input pixel format

* Create camera pipeline state

* Get YpCbCr to RGBA conversion shader working

* Rename XRVertex to Vertex

* Change static to constexpr for vertices

* Make CompileShader function static

* Remove typedef from Vertex struct

* Get drawing to output texture working in UpdateCameraTexture

* Redo changes undone in merge

* Release texture ref explicitly if creation from image fails

* Get YpCbCr to RGBA conversion shader working on macOS

* Rename local variable `bestPixelFormat` to `devicePixelFormat`

The "best" prefix in the `bestPixelFormat` local variable name only makes sense for the current iOS implementation. It doesn't make sense on macOS since we're not selecting a "best" device. On macOS we're selecting the default device.

This change updates the local variable name so it makes sense for the current implementation on both iOS and macOS.