Fix unit test failing for JSI engine
When using the JSI engine, the `Napi::Value::ToNumber` function does not coerce string values to numbers. This causes `setTimeout` to fail when the delay is given as a string.

This change fixes the issue by converting the string to a number using `std::stoi` instead of using `Napi::Value::ToNumber`.