Make timeout thread init after other members init
In some cases `m_thread` can be initialized and started before `m_mutex` is initialized, which causes an exception when the thread tries to lock the uninitialized mutex.

This change fixes the issue by initializing `m_thread` in the body of the `TimeoutDispatcher` constructor instead of in the initializer list. This ensures that `m_mutex` is initialized before `m_thread`.