Make `setTimeout` asynchronous and add `clearTimeout` (#1146)
The current implementation of `setTimeout` dispatches repeatedly until the given delay is reached. This causes performance issues.

This change improves performance by implementing `setTimeout` asynchronously on a separate thread.

The `clearTimeout` function is currently not implemented and since it is closely related to `setTimeout`, it makes sense to implement it in this change, too.