Fix `setTimeout` race condition (#1155)
In rare cases it's possible for the timeout function's shared pointer to call delete from the `setTimeout` worker thread instead of the JavaScript runtime thread. This happens when the timeout function is dispatched in `CallFunction` and called on the runtime thread before the worker thread removes it from the time and id maps. In this case, it is destroyed on the worker thread when the timeout is removed from the id map since the runtime thread has already released its shared pointer reference to the function. This causes a Napi assert since the function object has to be destroyed on the same thread it was created on.

This change fixes the issue by holding a shared pointer to the timeout function while the timeout is being removed from the maps in the worker thread, and then moving that held shared pointer to `CallFunction` to ensure the shared pointer ref count falls to zero on the JavaScript runtime thread.