Disable validation tests failing only on D3D12 (#1287)
Three validation tests for scissors are failing on D3D12 due to mip-maps
being created but not populated by bgfx. See
https://github.com/BabylonJS/BabylonNative/issues/1265. While we are
waiting for the fix in bgfx, we want to disable these tests on D3D12.

This change adds a key to the validation tests config.json named
`excludedGraphicsApis` that takes an array of strings specifying which
graphics apis to skip for the given test. For example, to skip the D3D12
graphics api when running one of the broken scissor tests, it would look
like this:
```
        {
            "title": "Scissor test",
            "renderCount": 10,
            "playgroundId": "#W7E7CF#12",
            "referenceImage": "scissor-test.png",
            "excludedGraphicsApis": ["D3D12"]
        },
```

The config.json key's name of `excludedGraphicsApis` was chosen to match
a similar key in the Babylon.js validation test config.json named
`excludedEngines`.