Remove Android xr double render (#1297)
This code has been in the Android XR implementation forever, and was
never understood. It did an explicit eglSwapBuffers to present to the
screen, which was needed in Babylon React Native as somehow the bgfx
rendering path was different between the two. Now, even in BRN this
results in double rendering. Maybe this is a result of one of the bgfx
updates. Since it seems to no longer be needed in any scenario, this PR
removes it. Unfortunately XR on Android in the BN Playground is broken
in a different way currently, so I have not been able to verify this
works correctly in that scenario. But previously this code was incorrect
in BN anyway, so I expect this is correct now in both BN and BRN.