Static engine allocator (#1327)
When shutting down the engine while having some textures, it's possible
for NativeEngine to be disposed before the textures.
In this lambda:


https://github.com/BabylonJS/BabylonNative/blob/95240a59f2e1483d4c4b92b329b84ff333f0de84/Plugins/NativeEngine/Source/NativeEngine.cpp#L216

The imageContainer has still a ref to NativeEngine::m_allocator that
gets invalid


https://github.com/bkaradzic/bimg/blob/5d61905eb1aa8074ee62553b4a7bd2dc16f4cf24/src/image.cpp#L3335

Setting m_allocator as static fixes the issue. It only contains a vtable
so it's threadsafe.