flipping happens after rendering (#1336)
bgfx default is to do page flip before rendering a new frame. So,
rendering 1 frame is not enough to see rendering on screen.
This 1 frame latency was visible on Apple and fixed with this PR:
https://github.com/BabylonJS/BabylonNative/pull/525
Unfortunately, the issue has been reported on Windows as well. With this
PR, flipping will always be after rendering.
Related posts: 
https://github.com/bkaradzic/bgfx/issues/339