Fix flash and image orientation for iOS image capture (#1363)
`ImageCapture` was implemented for iOS a while back, but had a few
outstanding bugs that are addressed in this PR:
1. Flash modes were not working - there were actually two bugs here, the
first being that we were trying to get supported flash modes from the
`AVPhotoCaptureOutput` before we actually created the
`AVPhotoCaptureOutput` (oops), and second there was a missing line to
increment the current index when populating the flash modes array.
2. There was no code to deal with image rotation, so the created image
was always sideways (I believe based on the orientation of the physical
camera sensor relative to the orientation of the device).

So the changes here are to fix the issues of bullet 1 and add code for
bullet 2.

I also made some small changes to experience.js to reflect the missing
support for invertY on Textures in BN which otherwise makes it confusing
why the code is the way it is.

EDIT: Regarding image rotation, I realized my fix for this was not
really correct. What actually happens with ImageCapture in the browser
(at least in Chrome on Android), which is similar to what camera apps
typically do, is that the image data itself is not rotated when the
photo is taken. Rather, it just includes correct Orientation metadata
(e.g. jpeg metadata), and code that reads that image is supposed to
respect the Orientation metadata. For example, the Photos app on iOS or
Windows reads this metadata when loading a jpeg and does the rotation
when displaying the image. Similarly, when a jpeg is used as a texture
in Babylon.js, the loading path relies on the browser respecting the
Orientation metadata when loading the image data. This means what really
needs to happen is:
1. We need to ensure the correct Orientation metadata is included in
photo capture (achieved by the new change that sets
`photoOutputConnection.videoOrientation`).
2. We need the Babylon Native image parsing code path to respect the
Orientation metadata, which bimg already gives us, we were just ignoring
it (new change that rotates the image data during image parsing if
needed).