Update to Babylon.js 7.6.2 (#1373)
Updating to Babylon.js version 7.0.0 reveals a whole slew of issues when
running the validation tests because
[#14868](https://github.com/BabylonJS/Babylon.js/pull/14868) broke the
render loop stopping behavior when running in native. Native does not
implement a way to cancel a requested animation frame. This missing
canceling functionality should be added to native someday, but, in the
meantime, [#15086](https://github.com/BabylonJS/Babylon.js/pull/15086)
will make it behave like it did before.

These changes fix potential issues in the code regardless of what
happened on the JS side.

- Update to latest JsRuntimeHost with lots of fixes.
- Call SetRenderResetCallback with null to clear the callback that can
cause a crash on shutdown.
- Update a few validation tests to not use render count.
- Update scripts to use `const`/`let` instead of `var`.
- Update validation script to handle some errors better and clean up
dead code.
- Add unhandled exception handler to Playground app so that validation
test will report correct errors and exit correctly.
- Change core graphics device implemenation to update bgfx state right
before requesting screenshots. This will ensure the screenshot is the
right size if the resolution has changed.
- Fix TestUtils:WritePNG to return an error when the byte length is an
unexpected size instead of silently failing.
- Change TestUtils::SetTitle to update the HWND title on a background
thread to avoid dead lock with the main thread.
- Update dynamic texture clip test to newest version that will fail if
the feature is not working.