Support non-normalized, non-float vertex buffers and unaligned vertex buffers (#1382)
This is a replacement for #1376. The original thinking was that we will
promote the non-float attributes to float which was a complex and poor
performance solution. The better solution is to change the shader to
take non-float attributes. Fortunately, this shader processing is
already happening for WebGPU on the JS side. @Popov72 [factored out the
shader processor](https://github.com/BabylonJS/Babylon.js/pull/15107) on
the JS side for WebGPU to also work with Babylon Native, which also
includes code that aligns the buffers to 4-byte boundaries. This native
PR ties everything together and removes all the promote to float logic
we had earlier to try to handle non-float attributes.

This change includes updating Babylon.js to 7.7.2. Using an older
version will result in a `console.error ` if non-normalized, non-float
vertex buffers are used.