Test installation iOS/Linux/MacOS (#1401)
forum:
https://forum.babylonjs.com/t/integrate-babylon-native-in-existing-swift-project/35886/28?u=cedric

- Because of CMake and Xcode, there are incompatilities with `install`
scheme on iOS. So, build needs to be started with the playground, then
using `cmake --install` . Also iOS install needs absolute prefix (
https://gitlab.kitware.com/cmake/cmake/-/issues/21282 )
- MacOS install scheme is fine but I could not enable scheme creation on
that target. So, it's been enabled for all targets.
- `if (TARGET ...` in cmake script test if the target exists. But if the
target exists but is not used, it will not be built. This is the case
for NativeTracing, disabling its build for iOS otherwise its .a is not
found at installation. Why is it not an issue for other build/installs?

- disable gtest fetching for ios
- disable openXR not other than win32 platforms
- `UNIX` is true in cmake scripts for Apple platforms, update checks
accordingly
- bump iOS deployment target to 15
- bump gradle task version because deprecate
- bump Xcode version from 15.3 to 15.4