Remove OpenXR dependencies (#1521)
The OpenXR dependency used by Windows builds needs to be updated to work
without errors when using CMake 4, but XR is not supported on Windows
anymore so there is no easy way to test the changes.

To fix this issue, this change removes XR support on Windows and removes
the OpenXR dependencies since they are only being used on Windows
(Android and iOS use different XR implementations).